доработка форм аутентификации
