--- src/pages/Login/styles.module.scss
@@ -1,57 +1,0 @@
-.container {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 20px;
-    height: calc(100vh - 110px); // 100vh - (height of header)
-  }
-
-.form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 40vw;
-    height: 65vh;
-    padding: 20px;
-    border-radius: 7px;
-    background: var(--black-15);
-}
-
-.heading {
-    text-align: center;
-    font-size: 30px;
-}
-
-.input {
-    background: var(--black-10);
-    border-radius: 3px;
-    text-align: center;
-    color: white;
-    width: 650px;
-    height: 80px;
-    margin-top: 8vh;
-}
-
-.signIn{
-    display: inline-block;
-    padding: 10px 15px;
-    font-size: var(--font-subheading-1);
-    border-radius: 5px;
-    color: black;
-    width: 450px;
-    background: var(--orange-50);
-    transition: var(--hover);
-    text-align: center;
-    margin: 5vh;
-    &:hover {
-        background: var(--orange-60);
-        transition: var(--hover);
-    }
-}
-.separator{
-    width: 650px;
-    height: 10px;
-    border-radius: 5px;
-    background: var(--black-10);
-    margin-bottom: 5vh;
-}
