+++ src/types.ts
@@ -1,4 +1,5 @@
 export enum RouteLinks {
     MAIN = '/',
     LOGIN = '/login',
+    REGISTRATION = '/registration',
 }
